[REFACTOR] introduce AttributionEntityTable
